Chapter 02 - Example 09: Dynamically allocated 2D array with multiple malloc call
Make multiple calls to "malloc", allocating an arrays. First, allocate a 1D array of N pointers to the element type, with a 1D array of pointers for each row in the 2D array. Then, allocate N 1D arrays of size M to store the set of column values for each row in the 2D array. Assign the addresses of each of these N arrays to the elements of the first array of N poitners.